Move `get_tuple_subtract_count`
